Added support for __slots__ and saving binary data to class variable
Added support for class attributes not included in `_field_order` by defining in `__slots__`

To simplyfy attribute search in `getattr` and `setattr` the metaclass now collects slots from self and any super class that is not `structObject`

Additionally, the binary data provided during `structObject.unpack()`
is now saved to `structObject._bindata`. This was added to allow
`structArray.len()` to compute it's length not explicetly stated
elsewhere in a data structure.